feat: refactor exception handling package structure and update imports
